requirements.txt
fastapi[all]
uvicorn
requests
python-dotenv
pymongo[srv]
stripe
yagmail